Fixed overflow issue in reverse and forward
Because STM32F103 uses 16 bit timers we overflow them alot with our high resolution encoders.  It creates a ticking sound in the motor when it overflows. It tries to reverse the motors since the overflow makes the computer think that we are going in the wrong direction. This however creates a bug which doesn't let us go from a positive speed to a negative. We can only go Pos -> 0 -> Neg or Neg -> 0 -> Pos. this has to be fixed.